Extract retryable_http_get to Downloader module
